feat: 사이드바에 다중 매장 처리
